divided getAllBlogs to utils directory
