+++ src/app/AussieLife/page.tsx
@@ -1,29 +1,11 @@
 import { Footer } from "@/components/Footer/page";
 import { Header } from "@/components/Header/page";
-import fs from "fs";
-import path from "path";
-import matter from "gray-matter";
-
-async function getAllBlogs() {
-  const postsDirectory = path.join(process.cwd(), "/src/posts");
-  const files = fs.readdirSync(postsDirectory);
-
-  const posts = files.map((filename) => {
-    const filePath = path.join(postsDirectory, filename);
-    const fileData = fs.readFileSync(filePath, "utf-8");
-    const { data, content } = matter(fileData);
-    return {
-      frontmatter: data,
-      slug: filename.replace(".md", ""),
-      content,
-    };
-  });
-  return posts;
-}
+import getAllBlogs from "@/utils/getAllBlogs";
+import ReactMarkdown from "react-markdown";
 
 export default async function AussieLife() {
   const posts = await getAllBlogs();
-  console.log(posts);
+
   return (
     <div className="flex flex-col min-h-screen">
       <Header />
@@ -33,7 +15,7 @@ export default async function AussieLife() {
           <div key={index}>
             <h2>{post.frontmatter.title}</h2>
             <p>{post.frontmatter.date}</p>
-            <p>{post.frontmatter.date}</p>
+            <ReactMarkdown>{post.content}</ReactMarkdown>
           </div>
         ))}
       </div>
